# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.LoginResponse:
    properties:
      message:
        type: string
      requires_2fa:
        type: boolean
      two_fa_token:
        type: string
      user_id:
        type: integer
    type: object
...
  models.ScanResult:
    properties:
      data:
        properties:
          attributes:
            properties:
              results:
                additionalProperties:
                  properties:
                    category:
                      type: string
                    result:
                      type: string
                  type: object
                type: object
              stats:
                properties:
                  malicious:
                    type: integer
                  suspicious:
                    type: integer
                  undetected:
                    type: integer
                type: object
              status:
                type: string
            type: object
          id:
            type: string
        type: object
    type: object
  models.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
  requests.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  requests.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  requests.Verify2FARequest:
    properties:
      token:
        type: string
      user_id:
        type: integer
    required:
    - token
    - user_id
    type: object
  responses.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Railgun AI-driven SIEM API
  title: Railgun API
  version: "1.0"
paths:
...
  /analytics/heatmap:
    get:
      description: Генерирует данные для построения карты интенсивности угроз
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Тепловая карта угроз
      tags:
      - Analytics
  /analytics/traffic/{hostId}:
    get:
      description: Получает статистику сетевого трафика для узла за временной период
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Трафик по конкретному хосту
      tags:
      - Analytics
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя по имени и паролю. Может вернуть токены
        сразу или потребовать 2FA.
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Если требуется 2FA
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "401":
          description: Неверные учетные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка генерации токенов
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход в систему
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Принимает Refresh Token и выдает новую пару Access/Refresh токенов
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Невалидный Refresh Token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить JWT токены
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создает новый аккаунт. Пароль должен быть не менее 8 символов.
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно создан
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Ошибка валидации или слабый пароль
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Имя пользователя уже занято
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - Auth
  /auth/verify-2fa:
    post:
      consumes:
      - application/json
      description: Проверяет временный код подтверждения и выдает финальные JWT токены
      parameters:
      - description: ID пользователя и код из приложения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT токены
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Неверный 2FA код
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Проверка 2FA кода
      tags:
      - Auth
...
  /integrations/scan:
    post:
      consumes:
      - multipart/form-data
      description: Загружает файл и отправляет его на проверку во внешние антивирусные
        сервисы
      parameters:
      - description: Бинарный файл для сканирования
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScanResult'
        "400":
          description: Файл не предоставлен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сканирования
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Сканировать файл через VirusTotal
      tags:
      - Integrations
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer <your_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"